+++ release-notes.md
@@ -3,6 +3,7 @@
 - Various bugfixes.
 - Add support fo ext-data-control-v1.
 - Implement wl-fixes.
+- Implement ei_touchscreen v2.
 
 # 1.7.0 (2024-10-25)